Cleaned_Type1_Allanson_Patricia.csv
I deleted a bunch of duplicates. The original source seems to be gone (?) so I wasn't able to check it against that. It also seems to me that, while she attempted murder a bunch of times, she never actually killed anybody, so she may need to be removed from the serial killer classification if we wind up adding in more types of crimes.